+++ setup_fp_env_system.sh
@@ -14,8 +14,8 @@
 #-----------------------------------------------------------------------------------------
 # Script information
 script_name='FP ENVIRONMENT - SYSTEM LIBRARIES'
-script_version="1.5.8"
-script_date='2020/10/26'
+script_version="1.5.9"
+script_date='2021/01/08'
 
 # Define file reference path according with https link(s)
 fileref_zlib='http://www.zlib.net/zlib-1.2.11.tar.gz'